Move spec helpers into `support` folder
